Move fixture up, test env doesn’t persist
