+++ korben/test/tier2/conftest.py
@@ -5,11 +5,19 @@ from korben.cdms_api.rest.api import CDMSRestApi
 
 @pytest.fixture(scope='session')
 def cdms_client():
+    'Placeholder for disconnect management and stuff?'
     client = CDMSRestApi()
     return client
 
-@pytest.fixture(scope='session')
-def staging_fixtures(cdms_client):
-    import ipdb;ipdb.set_trace()
-    # create
-    yield
+
+@pytest.yield_fixture
+def account_object(cdms_client):
+    'Create an AccountSet object and delete it aferwards'
+    create_resp = cdms_client.create('AccountSet', data={'Name': 'Timelort'})
+    assert create_resp.ok
+    acc = create_resp.json()['d']
+    yield acc
+    del_resp = cdms_client.delete(
+        'AccountSet', "guid'{0}'".format(acc['AccountId'])
+    )
+    assert del_resp.ok
